Use blockAction.when decorator to have a message reported in secure mode when trying to open the symbol dialog. (#13539)
Fix-up of #13535
Follow-up of #13500

Summary of the issue:
In #13500 a decorator has been introduced to speak a message when an action is unavailable in secure mode.
In parallel an NVDA 2021.3.5 patch release has been produced; this release contains a fix preventing to open the symbol dialog in secure mode. The 2021.3.5 (rc) branch has then been merged into master to get this fix in master.
This led to the fact that no message was reported when trying to use a script to open the symbol dialog in secure mode.

Description of how this pull request fixes the issue:
Added the blockAction.when decorator where it was missing.