Only remove XML tags if JAB widget's text starts with <html> tag (#13531)
Fixes #13102

Summary of the issue:
NVDA was incorrectly stripping text from Java widgets if the text looked like an XML tag. According to the Oracle Java docs, when using HTML in a Java widget you should start the text with the <html> tag.

Description of how this pull request fixes the issue:
This change gets NVDA to check if the text starts with <html> and only then will remove any text which appears to be XML tags.