Add secure mode decorators to future-proof code (#13548)
Summary of the issue:
Certain mainFrame commands `gui.mainFrame.ExampleCommand` are only exposed via the NVDA menu not in secure mode.
In the future, input gestures might be created for these commands, exposing them in secure mode.

Additionally, the log viewer is disable in secure mode, so script_navigatorObject_devInfo should be blocked, as it is misleading. The logs should not be accessible when running in secure mode.
Debug logging in secure screens can be configured with the serviceDebug parameter.

Description of how this pull request fixes the issue:
Adds secure mode decorators to commands that are currently inaccessible in secure mode.

Testing strategy:
The decorator has been tested in #13539 and other PRs.
Other than script_navigatorObject_devInfo, these messages should be currently inaccessible to the user.