Fix ARIA treegrid test to expect group position on row in focus mode (#13592)
Our system test for ARIA treegrid fails as it does not expect group position (E.g. 1. of 1) to be announced when focusing a row in focus mode.
It seems that although the test always exposed aria-posinset and aria-setsize, Chrome was not exposing this via accessibility until now (Likely Chromium 100).

Description of how this pull request fixes the issue:
Update the test to expect "1 of 1" at the end of speech.