Advanced settings panel: restore accelerator key (fix-up of #13437). (#13563) (#13590)
PR #13563 restores an accelerator key that has been removed by #13437 during NVDA 2022.1 release cycle.
Unfortunately, PR #13563 has been opened after translation freeze and thus could not integrate 2022.1.
However in #13563, I have written:
I would have wished to open this PR against beta to fix the issue immediately before the stable release. However, we are already in the translation freeze, thus this is not possible to change a translatable string, unfortunately. If a new translation freeze should occur, please reconsider merging this PR in beta.
According to this thread, an extension translation freeze is in preparation with new translatable strings.
Thus this allow to integrate this fix for 2022.1 considering that the accelerator key was removed during 2022.1 dev cycle.

Description of how this pull request fixes the issue:
Cherry-pick the commit corresponding to #13563, i.e. c446688.