Fix NVDA interaction with Java progress bars (#13594) (#13595)
fixes #13594

Summary of the issue:
Currently, NVDA does not beep on Java progress bars. This happens because the progress bar is not added to the overlay classes when Java sends the "progress bar" role to NVDA.

Description of how this pull request fixes the issue:
This pull request adds the missing ProgressBar class when it detects the "progress bar" role.