Fix line number reporting in Visual Studio (#13604)
Closes #13574

Summary of the issue:
In Visual Studio, when line number reporting is enabled in VS itself, line numbers are part of the text. Therefore they are reported regardless of NVDA's line reporting settings. This also causes indentation reporting to fail.

Description of how this pull request fixes the issue:
This pr isolates the line number from the text and ensures that it is reported appropriately when line number reporting in NVDA is on.

Testing strategy:
With line reporting enabled in Visual studio

Ensure that line numbers are reported when enabled in NVDA
Ensure that line numbers are not reported when disabled in NVDA