Fix HTML characters in mozilla rich edit causing error (#13644)
Link to issue number:
Closes #12542

Summary of the issue:
In mozilla applications when pasting certain HTML into rich edit controls it can lead to errors when navigating through the text. One such example is in the thunderbird compose message edit area.

Description of how this pull request fixes the issue:
The function _getEmbedded in ia2TextMozilla.py returns Optional, however the return value is never checked to be not None and this assumption leads to errors later on when the value is used. This pull request adds a check for None being returned from _getEmbedded and skips it if None is found.