Make POSITION_FIRST and POSITION_LAST relative to visible text in FORMATTED consoles.
The ability for users to explore all console text enabled by #12669 has been well appreciated, but it poses problems in paginated output (less, more, etc.) as the review cursor jump to top/bottom commands are relative to the entire buffer, which can grow quite large. To ease review of paginated output, make these commands jump relative to visible (not full) text. Note that it is now impossible to jump quickly to the top/bottom of the entire buffer.
Closes #13157.