Only close seika connection if initialized (#13684)
Closes part of #13679

Summary of the issue:
It appears that when resetting a config, the seika notetaker driver can be terminated before it has been initialized.

Description of how this pull request fixes the issue:
Only terminate the notetaker driver if it has been initialized.

Known issues with pull request:
Another issue is highlighted in #13679. While terminating/initializing during a config change, the braille.handler is attempted to be used when handling events on NVDA objects.
This is a more complex problem and needs a braille device to reproduce and test.