Announce readonly state for Java text controls (#13706)
Closes #13692

Summary of the issue:
In a Java swing application NVDA does not announce a read only text control as being read only.

Description of how this pull request fixes the issue:
Java Access Bridge does not apply the editable state to read only controls. So when mapping JAB states to NVDA states check whether the control is a text control and whether it lacks the editable state, applying the read only state accordingly.