Fix for navigator object highlighting activation (#13641)
Summary of the issue:
When activating Highlight of navigator object, the highlighting rectangle does not show up. It only shows up when the navigator object is moved for the first time. If the highlight of nav object is previously disabled, the issue occurs for all the four following types of activation:

When checking the corresponding checkbox in the vision parameters
When you reload NVDA's saved config provided that highlighter is enabled in NVDA's saved config.
When activating with a gesture a profile containing nav object highlight enabled
When NVDA starts up if the default config has nav object highlight enabled
Description of how this pull request fixes the issue:
Checks nav object initialization when the providers are initialized. Prior to this PR, only the focus initialization was checked (and the caret thanks to it).