Improve pronunciation of : * / in words
In contemporary German, there are multiple ways to indicate that a noun refers to all genders, e.g. users could be referred to as:

Nutzer_innen
NutzerInnen
Nutzer*innen
Nutzer:innen
Nutzer/innen
Nutzer/-innen

The common pronunciation is to use a glottal stop before the suffix. NVDA (with the default OneCore voices) does the right thing for _ and the capital I, but either uses a long pause (with : ), or reads out the character in default settings. This is distracting to the user when encountering texts using one of the other forms.

This patch fixes all the above cases to use a short pause, most closely resembling a glottal stop.