Installation of Add-ons
- The add-on store needs a way to allow installation.
- The actions should happen in the background, steps like downloading may take significant time.
- Progress should be discoverable.

Description of change:
- Introduces a way to associate actions with the GUI.
- To cater to power users a context menu has been added to list items (available add-ons)
- To ensure discoverability, buttons for actions are added to the details.

Note
There is currently an "update" action, which only logs.
It is included just as a demonstration / proof of concept.

There are many rough edges in this implementation, but since this is already a sizable change,
and many further changes are more likely to self contained it is a good checkpoint.

Testing strategy:
- Used the links in the add-on store data to download several *.nvda-addon files
- The manifests were modified (in various ways) to create different installation outcomes.
- The python "simpleHttpServer" module was used to srt.
- The links to the URL for add-ons in the cached addon-store file were pointed to localhost.
- Used 'Clumsy' tool to simulate slow connections, adding latency to the
  download process.

Known issues:
These will be worked on separately:

- The status of installed add-ons isn't reflected in the list, installed
add-ons with a matching version should be omitte". There is a problem
here, an add-on manifest does not include a numerical version, only a
version name string. Ordering based on version number with the add-on
store will not be possible for side-loaded add-ons.
- Pressing enter on an addon should open the context menu.
- Pressing enter on a link (in the details) should open the URL
- When closing the dialog (after inste blocked by add-ons being installed
(consider NVDA exit / os shutdown), and progress shown.
- Task management required for download / install steps. Install steps for
multiple add-ons should be serialized. Download can happen in parallel.
- Ok / Cancel should be eplaced wih a 'Done' button
- Add-ons optionally have an install_task which runs post installation, prior
to restart. This can be disruptive when installing multiple add-ons from the
store. Consider making this occur post restart.