add pronunciation for raw Unicode braille cells
If Unicode braille occurs within text — for instance, ⠐⠣⠃⠗⠇⠐⠜ says "(braille)" in UEB — NVDA currently just skips over the braille entirely. This makes it at least pronounce the braille cells.

Because this is pronunciation for the Unicode characters themselves, it is braille-code-agnostic and just pronounces individual braille characters in the usual "dot 3 5 7" way. If the input table is known, this default should get overridden, but it's at least a fallback, and allows one to read text that's _about_ braille.

I've left level & preserve in defaults.