For single column SysListViews, ignore column order array as it is not required (#13761)
Fixes #13659, #13735

Summary of the issue:
Certain applications don't provide a column order array. For the case of single column list views, a column order array is unnecessary, as a default mapping can be assumed.

Description of how this pull request fixes the issue:
For single column list views, map the first column in the UX, to the first column index: i.e. [0] is the column order array. When fetching from a multi column list view, log an error if the column order array is unknown.