Change input help documentation for move to first/last row scripts of fake table navigation in list views. (#13788)
Follow-up of #13435.

Summary of the issue:
In fake table navigation in list views, horizontal movements only move the navigator object, whereas vertical movements move the navigator object but also the focus. This was not reflected in input help for the newly introduced commands to jump to first/last row.

Description of how this pull request fixes the issue:
Updated the two scripts documentation.
While at it, I have also converted remaining fake table navigation scripts to syntax with decorator.