Add development approach to PR template (#13730)
Summary of the issue:

In order to aid reviewing complex pull requests, having a description of the development approach is helpful.
Often this is implicitly included in "Description of how this pull request fixes the issue", however this is also used as a high level summary of the user facing changes. A separation of concerns is needed.

Description of how this pull request fixes the issue:

Adds this information to the PR template and explanation.