Remove redundant "-:NOREFS" from SymStore cli options (#13439)
As per the documentation (https://docs.microsoft.com/en-us/windows-hardware/drivers/debugger/symstore-command-line-options) `-:NOREFS` should be used with `/p`. Reference pointer files are not included unless we use the `/p` flag. 

Description of how this pull request fixes the issue:

Removes `-:NOREFS`  and reorders options to be inline with documentation.