Fix-up of #13366: Replace dots with underscores before trying to check if the given App Module exists to work around bug in the Python import system (#13814)
Fix-up of #13366
Fixes #13813

Summary of the issue:
As part of PR #13366 I've modified the code responsible for retrieving application name from process id so that it no longer tries to import an App Module for a hosting process if it does not exist. Unfortunately Python find_Module method seems to have a bug - it returns a module when the provided string contains dots and the file named exactly like the last segment of the provided name exists in the package. For example if you have module named qq in the package, and you would like to check if the module named 6.5.qq exists True is returned even though there is no such module. I haven't yet tested how this behaves with more recent versions of Python nor opened an issue against them.

Description of user facing changes
NVDA no longer fails to work in applications containing dots in their file names in cases where you have an App Module for an application named exactly like the last segment of the application name.

Description of development approach
Dots are normalized to underscores before checking if the given module exists. While we did this already before trying to import a standard App Module this has not been done before trying to get module for a hosting process.