Fix-up changes for UIA by default
This commit fixes the link to the microsoft/terminal PR and adds a note to help users check their API level. I wasn't sure how to override the t2t preproc directive to write a raw hash symbol for the microsoft/terminal PR, so I didn't use it. Feel free to amend with an escaped hash.