Report ARIA details role (#13649)
Summary of the issue:
When announcing if an object has details, more information can be provided about the role of the details.
For example, if a comment is tied to a word, "has comment" should be reported instead of "has details".

Description of how this pull request fixes the issue:
Collects the details role.

The attribute details role is determined in a number of cases.

Braille
Use has cmmt to report presence of comment role aria-details target. Discussion

Focus Mode
detailsRole is normalized on the NVDAObject level to a controlType.Role.

Chromium
The IA2 attribute details-roles is used to get the role.

This supports the following roles (code): comment, definition, doc-endnote, doc-footnote. Other roles are given a value of * .

Currently only chromium supports details-roles.
Additionally, doc-endnote and definition are not supported IA2 attributes, and are not supported in NVDA widely.

We limit reporting for all cases by only reporting the fully supported details-roles: doc-footnote, comment

Other roles are reported as "has details".

FireFox
The related details node is fetched using in process injection, and the IAccessible information is retrieved.
The role is determined from the IAccessible role of the related node.

This is slower and provides a wider reporting of roles.

Browse Mode
The detailsRole attribute is added directly to the buffer, either as a descriptive role string or a role integer as a string.

Testing strategy:
The system test cases test the details role for Chrome, using both focus and browse mode.

runsystemtests.bat -t "*aria details*"

Equivalent manual testing was performed using those test cases on both Chrome and FireFox.