Added a margin to the browseable message. (#13885)
Summary of the issue:
The ui.browseableMessage has no margin. For people with visual difficulties, it makes more difficult to isolate the content of the browseable message from the text present in the background around the browseable message.

Description of user facing changes
The content of browseable message such as formatting information (double press of NVDA+F) is now rendered with a margin around the text.

Description of development approach
Just added the following style property to the message's body:
style="margin:1em"
The 1em measurement should correspond to 1 character.