Fallback to UIA if supported in Chromium and we don't have access to IA2 (#13032)
Summary of the issue:
Currently, UIA is used in the following cases in Chromium (notably Edge)

1. It is force enabled
2. We aren't able to inject in process

There is a particular case missing here. When running Edge under a different user, the IAccessible2 implementation is not available, whereas UIA works pretty nicely.

Description of how this pull request fixes the issue:
Add another check to ensure we use UIA in this case.