Make eSpeak's sonic boost be activated only when rate boost is enabled. (#13893)
A partial fix for issue #13876.

Summary of the issue:
With eSpeak, when rate boost is off, the speech rate is slightly higher (7% according to https://github.com/espeak-ng/espeak-ng/issues/131#issuecomment-263948128) when rate is 99% than when rate is 100%.

Description of user facing changes
With eSpeak, a speech rate of 100% corresponds to the fastest speech rate without rate boost; no more rate drop between 99% and 100%

Description of development approach
`_espeak.maxRate` is the rate used when NVDA speech rate is 100%.
So set `_espeak.maxRate` to 449 instead of 450. Indeed according to https://github.com/espeak-ng/espeak-ng/issues/131:
* the sonic library is used for eSpeak rate set to 450 or higher for many years (instead of 451 before).
* other ATs may already be using this new value, e.g. speech-dispatcher: https://github.com/brailcom/speechd (cf. https://github.com/brailcom/speechd/pull/603)

This allows to have a continuous increasing of speech rate when rate boost is off and to reserve the use of the sonic library for rate boost on.

If one day, eSpeak turns back to 450 for maxRate (what seems unlikely), we may revert this PR's content to follow eSpeak's decision.