Support non-interactive use of ensureVenv script (PR #13881)
Summary:
#13048 introduced another prompt to address when dependencies in requirements.txt have changed.
This blocks non-interactive use of the ensureVenv script, e.g. when running "scons source" non-interactively.

Description:
- Use defaults when running in non-interactive mode.
- Output to stdout notification when running in non-interactive mode.
  To aid debugging if interactive mode is determined incorrectly.
- Output questions and default answer to stdout when running in non-interactive mode
  To aid debugging if interactive mode is determined incorrectly.

Note: that isatty is not always available on sys.stdout 
e.g. when it is set to the NVDA python console.

Testing:
Tested running scons source > d:\output and observed that the venv was dropped and recreated.